// src/api/generated/models/clientResponse.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * BFF Service
 * Backend For Frontend service aggregating microservices
 * OpenAPI spec version: 1.0.0
 */
import type { ClientResponseVendedorAsignadoId } from "./clientResponseVendedorAsignadoId";

/**
 * Response schema for a client.
 */
export interface ClientResponse {
  cliente_id: string;
  cognito_user_id: string;
  email: string;
  telefono: string;
  nombre_institucion: string;
  tipo_institucion: string;
  nit: string;
  direccion: string;
  ciudad: string;
  pais: string;
  representante: string;
  vendedor_asignado_id: ClientResponseVendedorAsignadoId;
  created_at: string;
  updated_at: string;
}